feat: centralize authentication middleware
