+++ server/routes/auth.js
@@ -4,6 +4,7 @@ const { User, Client, UserClient, Invitation, Agency } = require('../models');
 const { generateApiKey, generateToken } = require('../utils/helpers');
 const { sendEmail, sendInvitationEmail, sendWelcomeEmail } = require('../utils/email');
 const supabase = require('../services/supabaseService');
+const authenticateToken = require('../middleware/auth');
 const router = express.Router();
 
 // Default permissions for client roles
@@ -29,24 +30,6 @@ const rolePermissions = {
     canViewAnalytics: true,
     canManageUsers: false
   }
-};
-
-// Middleware to verify auth token via Supabase
-const authenticateToken = async (req, res, next) => {
-  const authHeader = req.headers['authorization'];
-  const token = authHeader && authHeader.split(' ')[1];
-
-  if (!token) {
-    return res.status(401).json({ error: 'Toegangstoken vereist' });
-  }
-
-  const { data, error } = await supabase.auth.getUser(token);
-  if (error) {
-    return res.status(403).json({ error: 'Ongeldig of verlopen token' });
-  }
-
-  req.user = { userId: data.user.id, email: data.user.email };
-  next();
 };
 
 // Register new user
@@ -315,7 +298,7 @@ router.post('/invite', authenticateToken, [
 
     // Check inviter has admin access to client
     const inviterAccess = await UserClient.findOne({
-      where: { userId: req.user.userId, clientId, role: 'admin' }
+      where: { userId: req.user.id, clientId, role: 'admin' }
     });
 
     if (!inviterAccess) {
@@ -327,7 +310,7 @@ router.post('/invite', authenticateToken, [
       return res.status(404).json({ error: 'Client not found' });
     }
 
-    const inviterUser = await User.findByPk(req.user.userId);
+    const inviterUser = await User.findByPk(req.user.id);
     const frontendUrl = process.env.FRONTEND_URL || req.get('origin') || 'http://localhost:3000';
 
     // Create invitation token
@@ -339,7 +322,7 @@ router.post('/invite', authenticateToken, [
       token,
       role,
       clientId,
-      invitedBy: req.user.userId,
+      invitedBy: req.user.id,
       expiresAt
     });
 
@@ -465,7 +448,7 @@ router.post('/accept-invitation', [
 // Get current user
 router.get('/me', authenticateToken, async (req, res) => {
   try {
-    const user = await User.findByPk(req.user.userId, {
+    const user = await User.findByPk(req.user.id, {
       attributes: ['id', 'email', 'name', 'role', 'lastLogin'],
       include: [
         { model: Agency, attributes: ['id', 'name', 'settings'] },
@@ -503,6 +486,4 @@ router.post('/logout', authenticateToken, (req, res) => {
   res.json({ message: 'Logged out successfully' });
 });
 
-// Export both router and middleware
-router.authenticateToken = authenticateToken;
 module.exports = router;
